requirements.txt
elasticsearch
flask
pyyaml
python-dateutil